--- app/services/ripple.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-paper/services/ripple';
