+++ tests/dummy/app/templates/button.hbs
@@ -27,7 +27,7 @@
     {{#paper-button raised=true primary=true action="targetButton" target=component bubbles=false}}Button no bubble{{/paper-button}}
   </p>
   <p>
-    {{#paper-button icon-button=true}}{{paper-icon "more-vert"}}{{/paper-button}}
+    {{#paper-button icon-button=true}}{{paper-icon "accessibility"}}{{/paper-button}}
   </p>
   {{/custom-button}}
   <p>
