--- tests/unit/services/ripple-test.js
@@ -1,12 +1,0 @@
-import { moduleFor, test } from 'ember-qunit';
-
-moduleFor('service:ripple', 'Unit | Service | ripple', {
-  // Specify the other units that are required for this test.
-  // needs: ['service:foo']
-});
-
-// Replace this with your real tests.
-test('it exists', function(assert) {
-  let service = this.subject();
-  assert.ok(service);
-});
